Let a caller act as a unit in their own subtree, and lock finalized vouchers
Phase 37 — دسترسی سلسله‌مراتبی به واحدها. A deliberate reversal, at the
project owner's request, of the exact-equality scoping built in phases
19/32/33: a user may now act as any unit inside their own subtree. What did
not change is who decides — the client asks, the server validates.

  * Three read-only endpoints: GET /api/me, /api/me/accessible-units,
    /api/years. None takes a subject parameter, which is the security
    property rather than shorthand: a route accepting a unit code would be
    an enumeration surface.
  * The chosen unit travels as the X-Vahed-Code header, not a body or query
    field, so it applies to all ~60 existing commands and queries without
    changing a single contract and never enters a request model.
  * IUnitScopeResolver is the only place untrusted input becomes a scope.
    VahedOwnership is untouched: because the behavior stamps the effective
    unit, record-level ownership keeps working unchanged.
  * The ستاد rule (TYPECODE 17) is ported as a named constant. Live Oracle
    confirms exactly one unit carries it, and that ستاد مرکزی has only 7
    direct children of 1046 — it reaches everything through that rule, not
    through the tree. The tree itself is two levels deep with no
    grandchildren, so recursive and one-level descent agree on today's data.

Three bugs surfaced while testing this, all fixed:

  * TB_YEAR.WORKING_YEAR was scaffolded as byte against a NUMBER(4) column,
    so every read of a real row (1405, 1404) threw from GetByte. Present
    since phase 2 and invisible because nothing had queried the table. A
    convention test now sweeps the model; it found 29 more such columns,
    latent rather than active on live data, recorded in a documented
    allowlist instead of being quietly left.
  * ORA-00904: "FALSE" — projecting a boolean comparison. Oracle has no
    boolean type. The rule now recorded: select the value, decide in C#.
  * The repository test suite runs on SQLite and is blind to that entire
    class of bug — the second time it has cost us. OracleBooleanProjection
    Tests generate real Oracle SQL offline, and one test reproduces the
    original defect so the guard cannot pass for the wrong reason.

Phase 38 — قفل ویرایش/حذف سند. A voucher is editable only in یادداشت and
موقت; بررسی‌شده and تأیید دائم are view-only and must be moved back with the
separate change-state operation. VoucherEditability is the single decision
point and runs on all five write paths, lines included — locking the head
alone would leave a finalized voucher editable through its own rows. 409
rather than 403: the caller is entitled to edit it, just not in this state,
and the message says how to get there. An absent or out-of-enum DOCLIFE
fails closed, since the column's DEFAULT 0 has no known meaning.

2750 tests green.

+++ backend/src/Accounting.Domain/Entity/TB_YEAR.cs
@@ -5,7 +5,17 @@ namespace Accounting.Domain.Entity;
 
 public partial class TB_YEAR
 {
-    public byte WORKING_YEAR { get; set; }
+    /// <summary>
+    /// سال مالی — Jalali year, e.g. 1405. Physical column is <c>NUMBER(4,0)</c>.
+    ///
+    /// ⚠️ <b>Was scaffolded as <c>byte</c>, which is wrong and threw at runtime.</b> A byte tops
+    /// out at 255, so <c>OracleDataReader.GetByte</c> raised
+    /// <c>InvalidCastException: Specified cast is not valid</c> for every real row (verified on
+    /// live Oracle 2026-09-22: rows 1405 and 1404). The bug survived unnoticed from the phase-2
+    /// scaffold until phase 37 simply because nothing had ever queried this table. Corrected to
+    /// <c>short</c>, which covers the full NUMBER(4) domain.
+    /// </summary>
+    public short WORKING_YEAR { get; set; }
 
     public bool? ISCURRENT { get; set; }
